Move SearchConditionExporterTestCase to Test namespace
